Reduce to one mailer for all
