+++ resources/locales/fr_FR/cake.pot
@@ -4,8 +4,8 @@
 #, fuzzy
 msgid ""
 msgstr ""
-"Project-Id-Version: CakePHP 4.4.7\n"
-"POT-Creation-Date: 2023-04-18 18:03+0000\n"
+"Project-Id-Version: CakePHP 4.4.15\n"
+"POT-Creation-Date: 2023-08-10 10:55+0000\n"
 "PO-Revision-Date: YYYY-mm-DD HH:MM+ZZZZ\n"
 "Last-Translator: NAME <EMAIL@ADDRESS>\n"
 "Language-Team: LANGUAGE <EMAIL@ADDRESS>\n"
@@ -259,20 +259,20 @@ msgstr ""
 msgid "The count does not match {0}{1}"
 msgstr ""
 
-#: ./vendor/cakephp/cakephp/src/Utility/Text.php:923
+#: ./vendor/cakephp/cakephp/src/Utility/Text.php:916
 msgid "and"
 msgstr ""
 
-#: ./vendor/cakephp/cakephp/src/Validation/Validator.php:2539
+#: ./vendor/cakephp/cakephp/src/Validation/Validator.php:2543
 msgid "This field is required"
 msgstr ""
 
-#: ./vendor/cakephp/cakephp/src/Validation/Validator.php:2559
+#: ./vendor/cakephp/cakephp/src/Validation/Validator.php:2563
 #: ./vendor/cakephp/cakephp/src/View/Form/ArrayContext.php:249
 msgid "This field cannot be left empty"
 msgstr ""
 
-#: ./vendor/cakephp/cakephp/src/Validation/Validator.php:2711
+#: ./vendor/cakephp/cakephp/src/Validation/Validator.php:2715
 msgid "The provided value is invalid"
 msgstr ""
