+++ resources/locales/fr_FR/default.pot
@@ -5,7 +5,7 @@
 msgid ""
 msgstr ""
 "Project-Id-Version: PROJECT VERSION\n"
-"POT-Creation-Date: 2023-08-09 17:54+0000\n"
+"POT-Creation-Date: 2023-08-10 11:44+0000\n"
 "PO-Revision-Date: YYYY-mm-DD HH:MM+ZZZZ\n"
 "Last-Translator: NAME <EMAIL@ADDRESS>\n"
 "Language-Team: LANGUAGE <EMAIL@ADDRESS>\n"
@@ -1441,10 +1441,6 @@ msgstr ""
 msgid "The email was sent to the user."
 msgstr ""
 
-#: ./src/Mailer/LordMailer.php:20
-msgid "Message Sent Through the LORD Contact Form"
-msgstr ""
-
 #: ./src/Mailer/UserMailer.php:17
 msgid "Reset your Password"
 msgstr ""
@@ -1461,8 +1457,8 @@ msgstr ""
 msgid "A Staff Member Sent You a Message"
 msgstr ""
 
-#: ./src/Mailer/UserMailer.php:81
-msgid "Message Sent Through the Contact Form"
+#: ./src/Mailer/UserMailer.php:82
+msgid "Message Sent Through the LORD Contact Form"
 msgstr ""
 
 #: ./src/Model/Behavior/PictureBehavior.php:84
@@ -7549,7 +7545,7 @@ msgid "<p>Hi, {0} !</p>\n\n<p>You have modified your registration email. Please 
 msgstr ""
 
 #: ./templates/email/html/contact_email.php:3
-msgid "<p>{0} sent the following message through the LORD contact form:</p>\n\n<p>“{1}”</p>\n\n"
+msgid "<p>Someone ({0}) sent the following message through the LORD contact form:</p>\n\n<p>“{1}”</p>\n\n"
 msgstr ""
 
 #: ./templates/email/html/reset_password.php:3
@@ -7569,7 +7565,7 @@ msgid "Hi, {0} !\n\nYou have modified your registration email. Please copy and p
 msgstr ""
 
 #: ./templates/email/text/contact_email.php:3
-msgid "{0} sent the following message through the LORD contact form:\n\n“{1}”\n\n"
+msgid "Someone ({0}) sent the following message through the LORD contact form:\n\n“{1}”\n\n"
 msgstr ""
 
 #: ./templates/email/text/reset_password.php:3
