+++ templates/email/html/contact_email.php
@@ -1,6 +1,6 @@
 <?=
 
-__('<p>{0} sent the following message through the LORD contact form:</p>
+__('<p>Someone ({0}) sent the following message through the LORD contact form:</p>
 
 <p>“{1}”</p>
